Simplify rollback arguments
Make the rollback command speak in terms of deployment configs and
versions, instead of deployments which aren't formal API resources.

This keeps the command consistent with other deployment related APIs
and commands.

Rollback now accepts a deployment config and version for the rollback
target; if no version is supplied, the last completed deployment will
be inferred.