Implement authenticated image import
Add a new API which allows images to be imported through an image
stream, with the ability to optionally tag those images into a new or
existing image stream. Allow the image import controller to continue to
work based on annotation.

Add the concept of "generation" to image streams to concretely record
the state at which a tag was changed, and allow importers to track when
that state was effected (into status). Add a conditions field to each
status tag to report failures to users incrementally.

Preserves all fundamental behaviors for legacy clients. Omitting
generation on an update results in the field being preserved. Clients
must set generation to 0 to request a clear. Clients are prohibited from
providing their own generation.

Update oc tag to not clear the top level annotation when setting the
direct field. Update oc import-image to prefer using the new method
to import. Update oc new-app to use the new import method, and only
set a single tag if the server is detected to support tag import.

Update image creation and update to extract layer names and size
information into the serialized image object.