Enforce --tty=false flag for 'oc debug'
Changes behavior for "oc debug" command, specifically with "--tty" flag. Prior behavior left "--tty" essentially useless, as it was not checked if the user specifically requested to disable TTY by passing "--tty=false" (if the command was run in a terminal, and "--no-tty" was not set to "true", it would always default to a TTY, even with "--tty=false" set.

This change allows both flags to remain for compliance with any other programs that might use them, at least until one is deprecated as it has been suggested that having both is redundant.

Fixes https://bugzilla.redhat.com/show_bug.cgi?id=1343358