Router sets its host name in admitted routes
The router is managed by the cluster admin and the admin knows the
external host name of the router. The user creates routes to route
desired host DNS names to the the application. The user doesn't have
access to the router (which is usually in a different namespace). When
a router admits a route the user needs to work with the external
network admin to point the route's hostname to the router. Previously
the user had to contact the cluster admin to find the host name of the
router.

This change adds the --router-canonical-hostname= option to oadm router
which creates and populates the ROUTER_CANONICAL_HOSTNAME environment
variable in the router's dc. This is a string that contains the
canonical hostname of the router. Whenever the router reloads and
evaluates the routes the route's status is edited to include the router's
host name so the user can quickly access the information. If the
ROUTER_CANONICAL_HOSTNAME is updated or added to an existing router the
new information is updated in the admitted routes on the next reload.

The route (oc get route <routename> -o yaml) status reports the name
of the router(s) that have admitted the route. This change adds the
Router's host name to the status in the route.

The information is also reported in "oc describe route" as shown below.

Example:

oadm router myroutername --router-canonical-hostname=oh.my.heavens.org

oc get dc/myroutername -o yaml
...
spec:
  template:
    spec:
      containers:
      - env:
        - name: ROUTER_CANONICAL_HOSTNAME
          value: oh.my.heavens.org

oc get route hello-route3 -o yaml
...
status:
  ingress:
  - conditions:
    - lastTransitionTime: 2016-12-07T15:20:57Z
      status: "True"
      type: Admitted
    host: hello-openshift-v3.not.in3.mycloud.com
    routerCanonicalHostname: oh.my.heavens.org
    routerName: myroutername
    wildcardPolicy: None

oc describe route/hello-route3
...
Requested Host: hello-openshift-v3.not.in3.mycloud.com
          exposed on router myroutername (host oh.my.heavens.org) 12
minutes ago

bug 1393489
https://bugzilla.redhat.com/show_bug.cgi?id=1393489
Resolves origin-web-console issue 336
           https://github.com/openshift/origin-web-console/issues/336
openshift-docs PR#3377

Signed-off-by: Phil Cameron <pcameron@redhat.com>