Allow objects to request resolution, not just images
A lot of times, you want to be able to reuse a bit of config that works
fine elsewhere. By adding the `alpha.image.policy.openshift.io/resolve-names`
annotation with value `*` to the object you can request that resolution
attempt to match all image references to an image stream tag in the
current namespace, regardless of their own policy. This makes it easy to
transition to looking up image stream tags by name.

Extend `oc set image-lookup` to accept `deploy/foo` and to be able to
list and display non image stream resources.