Add oadm migrate volumesource
This command confirms that all objects that contain
volumeSource.metadata do not have it set.  volumeSource.metadata is
deprecated and will no longer be available in the next release.  Thus
all objects that contain a non-nil volumeSource.metadata must be
manually converted to prevent data loss.

The following resource types are checked by this command:

* cronjobs.batch
* daemonsets.extensions
* deploymentconfigs.apps.openshift.io
* deployments.extensions
* jobs.batch
* podpresets.settings.k8s.io
* pods
* podtemplates
* replicasets.extensions
* replicationcontrollers
* statefulsets.apps

No resources are mutated.

Signed-off-by: Monis Khan <mkhan@redhat.com>