Deployer pod may be unable to observe started pod from API
The code in the deployer pod may reach the lifecycle hook execution
prior to the kubelet being able to update the pod status to "running"
with a valid start time. This caused the deployer pod to panic because
start time was empty.

Since the window for this race is small, if the deployer pod can't see a
start time it can use `time.Now()` to bound the expiration window. Since
the process is reentrant, we can't use a start time at the beginning of
the deployer process (it's no more accurate that way).

A future change should guarantee the start time is available to the
container (via downward API, probably).