Not found errors must match object in migration
This change tightens the code to not tolerate not found errors that
refer to a different object than what was being migrated.  For
example, the previous code would not error if the not found error
refereed to the namespace being missing instead of the actual
object.  Thus we can detect objects that have been orphaned inside
of a deleted namespace.

Signed-off-by: Monis Khan <mkhan@redhat.com>