Add PKCE support to oc
This change updates RequestToken to support an OAuth code flow along
with the existing token flow.  The code flow is set as the default.
RequestToken also performs a check against the server's OAuth
metadata endpoint.  If the server supports PKCE with S256 (which
all recent versions of OpenShift do), than RequestToken includes
PKCE data to strengthen the code flow since we must use a public
OAuth client with oc.  Care was taken to preserve the existing
behavior of following redirects that are part of the challenge flow
(an authenticating proxy for example) while acting on redirects that
are instead part of the OAuth flow (exchanging the code for an
access token).  The use of all OAuth URL constants was consolidated
into a single location to guarantee they are used correctly.

Signed-off-by: Monis Khan <mkhan@redhat.com>