prevent client from looking up specific node
PR #16387 prevented a debug pod from being created at all, if a
specified node-name pointed to a node that did not exist.

Although the patch aimed to make "debug" command failures more
immediate, a client should not attempt to retrieve a specific node in
the "debug" command, as there is no guarantee that the user has access
to retrieve node information. While this means that the "debug" command