Correctly handle NotFound errors during migration
This change makes it so that the migrate storage command will now
reprocess the body of a failed GET request to follow the standard
path of extracting a status error.  This normalizes the structure of
the errors the code must handle.

When comparing the resource.Info to an error, the info's REST
mapping is now used to extract group kind information since the
associated runtime.Object is not guaranteed to have valid type meta
data.  This is combined with relaxed case-insensitive matching to
make sure that we only fail the migrate storage command on NotFound
errors that we know do not match the info.

When a NotFound error occurs, the migrate command now correctly
reports that it did not change the resource.  Previously it reported
that it had successfully migrated the resource, which is not true
since it is impossible to migrate a resource that does not exist.

Bug 1506006

Signed-off-by: Monis Khan <mkhan@redhat.com>