Parallelize storage migration
Changes made to "oc adm migrate storage":

1. Use a channel based producer / consumer pattern to parallelize
the underlying migrator logic.
2. Unit test new parallel code with the race detector.
3. Chunk list requests to 500.
4. Add a bandwidth flag to control the rate of network IO in Mbps.
This defaults to 10 Mbps and can be set to zero to mean "no limit."
5. Deprecate and hide the confirm and output flags.  This removes
the "dry run" mode from this command which never made any sense and
was almost certainly always invoked as a mistake.  Old scripts can
continue to set these flags; the value will simply be ignored and a
deprecation warning will be printed (but the command will not fail).

Signed-off-by: Monis Khan <mkhan@redhat.com>