Support login to the integrated registry via CLI
Adds a new `oc registry login` command that allows the user to easily log
in to the integrated registry. By default it tries to use the current
user token (but supports `-z` for service accounts) and looks up the
public or internal hostname by finding image streams in the current
namespace or the 'openshift' namespace (which is usually populated).

Also add `oc registry info` for fetching the current registry DNS name.