Replace isNotFoundForInfo with errors.IsNotFound in migrate logic
This change removes the custom isNotFoundForInfo in favor of a
simple errors.IsNotFound check.  In the past it was possible to
check that a NotFound error was for a specific object (to aid in
locating subtle bugs in the API server that we have encountered in
the past).  However, upstream changes [1] have made it so that if
you try to update a non-existent namespaced object in a non-existent
namespace, resource.Info.Get responds with a NotFound for the
namespace and not the object itself.  At this point it no longer
makes sense to maintain the fragile logic that isNotFoundForInfo was
using.

[1] kubernetes/kubernetes#31594

Signed-off-by: Monis Khan <mkhan@redhat.com>