Add a new `oc adm release mirror` command
This will enable an administrator to mirror an online release payload
to a local registry in preparation for running offline. It gathers the
images referenced by a release payload, mirrors them to the destination
registry, and then prints out the target payload image reference and
the prefix the CVO will use for offline.

    $ oc adm release mirror --from=registry.svc.ci.openshift.org/openshift/origin-release:v4.0 \
         --to registry.svc.ci.openshift.org/clayton-test-2/test