Use group resource in migrate reporting
This change updates the reporting during migration to use the group
resource instead of just the resource.

Thus the confusing messages with "the server does not allow this
method on the requested resource" for pod metrics will go from:

-n hasha-pro1 pods/nodejs-ex-2-kmzrv

to:

-n hasha-pro1 pods.metrics.k8s.io/nodejs-ex-2-kmzrv

This makes it clear what resource is being referenced.

Bug 1631517

Signed-off-by: Monis Khan <mkhan@redhat.com>