Generate cincinnati metadata when building a release
By default, no metadata is added since a release payload should not require it.
If user specifies `--previous`, `--metadata`, or `--release-manifest` we validate the
inputs and generate manifest as `cincinnati` (naming under discussion) in the resulting
`/release-manifests` directory of the output image.

Change `--name` to default to `0.0.1-<date>` so that the default name is semantic.

Refactor the Run method to be smaller.

Peek at the contents of the release zip so that during image pushes we don't print
the error (most of the time) while pushing the layer.