Add `--image-for` and `--output` to `oc adm release info`
`--image-for=TAG` lets a user get the exact pull spec out of the
payload or return an error. Allows scripting around a particular
payload, for instance to get the installer or CLI binary out of
the image referenced in the payload.

The `--output` command acts like `oc image info` and prints a
semi-formal API output of the info gathered for scripting.