Make the release image vary only by inputs
When generating the tar layer, ensure the timestamps of content is
predictable based on the input in order to guarantee two users
generating a payload from the same inputs will get an identical tar
(within the bounds of golang's tar/gzip implementations).

Remove some old code around reading payloads from disk that can't
be used now that we have to replace versions from the originals.