Extract openshift-install and use the exact SHA
Ensures that the version the tool is created with is always used by
default.

Add a new `release.txt` file to the release output that contains info
about the release and a brief description indicating that the installer
is locked. Add a pull spec to `oc adm release info` by default.

Add `--signing-key=PRIVATEKEY` to `extract --tools` to support simple
signing of the sha256sum.txt file to create integrity on release. This
support will remain simple and more complex use cases can be done with
gpg.