pkg/cli/admin/release: extract to target command not source
When extracting commands from a release, the command is typically the
same as the source executable in the image, however if that is not the
case - for example, in the case of the baremetal installer, oc does the
wrong thing.  If the command is `openshift-baremetal-installer` but it
is extracted from `usr/bin/openshift-installer`, oc extracts the file as
openshift-installer.