Introduce MOTD for login command
Due to legal requirements, some deployments need to display a legal
notice for users "entering" a system.

This legal notice can range from telling the user that they are entering
a government system and are being monitored, to just displaying the
rules of usage in the system. We'll implement this requirements in the
form of a "message of the day" (MOTD).

This is required by law by several countries, with the intent that in
case the user attacks the system and needs to be prosecuted, the message
can be used as proof that the attacker knew against whom the attack was being
executed.

In this PR, the "oc login" command will search for the motd
ConfigMap, which would ideally be configured in the "openshift"
namespace.

If the notice is not present, or the user has no rights to access that
ConfigMap, the command will not fail.

Note that the legal notice will be displayed for the login command
regardless if the user is authenticated or not.

As a separate PR, the (empty) ConfigMap and the needed ClusterRole and
ClusterRoleBinding will be populated by default in the deployment.