pkg/cli/image/extract: disable pigz to prevent race condition
There is a race condition in the vendored version of docker code that
can cause image extraction to panic.  When the `pigz` package is
installed, docker's DecompressStream code prefers to use it, however,
that code can return the io buffer to the pool while the command is
still writing to it. If the buffer is reused while that's happening,
`oc` will panic.  There are many reports of this happening.

As the vendored docker comes from kubectl, and the version there (even
for k8s 1.16) is quite old, this disables using pigz at all by setting
MOBY_DISABLE_PIGZ environment variable.