upgrade: Separate flags for safety instead of abusing force
The --force flag is dangerous and potentially allows untrusted
content to be upgraded to accidentally. Instead, introduce two
new flags `--allow-explicit-upgrade` (for upgrading to something not
in availableVersions) and `--allow-upgrade-with-warnigns` (for upgrading
when another upgrade is in progress or the cluster is reporting
an error) and remove those checks from `--force`.

While this is an API change, it is necessary to ensure that users
do not accidentally get access to untrusted content when
performing upgrades across major versions in advance of graph
updates, or when they are upgrading in disconnected environments.