Display MOTD on interactive actions (debug, exec, rsh, run, attach)
Whenever one of the aforementioned commands is executed, and the command
is executed on interactive mode, the MOTD will be displayed. This is
similar to the behavior that SSH provides, where running a single
command on a node will not trigger this behavior, but initiating a
session will.

This work is done to address compliance requirements.