registry-login: Allow login with credentials
We have no way to login to an arbitrary registry, like quay.io. This
isn't terribly common, but feels like a gap in a command called
`registry login`. This allows `--basic` to provide `user:password`
instead of service account or the user's ID. You can now run:

    oc registry login --to=/tmp/config --registry=quay.io --auth-basic foo:bar

The command now outputs a message if successful.