cmd/oc: disable docker's use of pigz earlier
Due to a race condition in Docker[1], we need to disable extraction
using unpigz. Currently this is done in the image extraction code,
however this code is multi-threaded.  Setenv is not thread safe in C[2],
so even thought it is safe in go[3], there's a small risk if there's
any C threads running.

It's safer to just set this env variable once when `oc` starts, rather
every time the layerByEntry function is called.

[1] https://github.com/moby/moby/issues/39859
[2] https://github.com/rust-lang/rust/pull/24741
[3] https://github.com/golang/go/blob/a38a917aee626a9b9d5ce2b93964f586bf759ea0/src/syscall/env_unix.go#L18