pkg/cli/admin/upgrade: allow users to force when updating to latest
--force is used when
a) the release image is not verified.
b) the operators are blocking upgrades.

For cases when the users are using custom update channel, the release-image is probably not going to be verified by Red Hat signature, but the users should be allowed to update to the latest
available update by skipping the verification step.

And for cases where the users want to upgrade to the latest upgrade, even when the operators are blocking updates should be achievable by setting the force flag.