Enable all Linux arches in cli-artifacts
In order for console-operator deployment to succeed, cli-artifacts needs
to be available on all arches for downloads-openshift-console.  However,
in that case, /usr/bin/oc (inherited from cli) is a native binary, and
we want to provide all primary Linux architectures to match those on
mirror.openshift.com.

In order to do so, this provides cross-compiled Linux binaries for
multiple architectures.  Cross-compiling oc fails with gssapi enabled,
therefore it is disabled in the cross builds.