pkg/cli/admin/release/extract_tools: Drop AsArchive from extractTarget
This is a per-extraction-call setting, not a per-extracted-target
setting.  Targets for a given extraction call are either going to be
all archives or never archives.  AsArchive landed in
openshift/origin@a63351b9bb (release: Support extracting the 'oc' and
'openshift-install' binaries, 2019-03-30, openshift/origin#22439).

Dropping the AsZip bit of the loop is fine because we are hard-coding
AsZip true for the windows oc in the initial availableTargets
declaration.