pkg/cli/admin/release/extract_tools: Support old paths
These are where the binaries lived before e949088d (Enable all Linux
arches in cli-artifacts, 2019-11-07, #153), and since the extraction
chokes on hardlinks [1], we need to be looking at those locations to
extract from 4.2.z and older releases.

Also shuffles the Dockerfile around to move the darwin and Windows
builds to their old locations and symlink them from the standard
locations.  That way older 4.2.z releases will find regular files at
the locations they expect, but that we'll be able to serve files from
the standardized pattern in the downloads deployment [2].

I'm also removing the cli-artifacts-built linux/amd64 binary in favor
of the build we inherited from the parent cli image.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1774642
[2]: https://github.com/openshift/console-operator/pull/354