pkg/cli/admin/release/extract_tools: Make some missing targets nonfatal
With this command, we still error out if we didn't extract anything.
But if we got through the tarball without erroring on an extraction
and we got something out, just accept that missing files are because
the image being used didn't have some possible target, and that that's
ok.  Avoids:

  $ oc adm release extract --tools --filter-by-os='.*' quay.io/openshift-release-dev/ocp-release:4.2.4
  error: unable to find multiple files: usr/share/openshift/linux_arm64/oc, usr/share/openshift/linux_ppc64le/oc, usr/share/openshift/linux_s390x/oc

when we expect 4.2.4's cli-artifacts to be missing those new non-amd64
binaries.