debug: update support tools container image
The RHEL7 support tools image is only built for AMD64, but there will
soon be a need for an image that works on System Z (and eventually
Power). Changing to the RHEL8 image gets us the additional
architectures. Additionally, RHCOS has moved to RHEL8 content, so it's
best to keep the support tools in sync.