release: Add a new annotation for component display labels
We want to report a "pretty name" for RHCoS and FCoS that can't be
inferred from the component name (machine-os). Add a new supported
label to components that allows them to set a display name for
the component. The `io.openshift.build.version-display-names` label
behaves like versions, but takes NAME=DISPLAYNAME and the changelog
is updated to show the display name in preference.