Bug 1800674: Option to force RepoDigest mirror
Currently `oc image mirror`, when pointed at a manifest list digest,
checks the contents of the manifest and mirrors just the
underlying image if it contains only a single image.

This change adds an option to mirroring to force the repodigest to
mirror in addition to the underlying image digest. Additionally, it
forces `oc adm catalog mirror` to always enable this option since
the repodigest is what is encoded into the OLM manifests.

https://bugzilla.redhat.com/show_bug.cgi?id=1800674