Bug 1804463: Option to force RepoDigest mirror
Currently `oc image mirror`, when pointed at a
manifest list digest, checks the contents of the manifest and mirrors
just the underlying image if it contains only a single image.

This change adds an option to mirroring to force the repodigest to
mirror in addition to the underlying image digest. Additionally, it
forces `oc image mirror` and `oc adm catalog mirror` to always
enable this option

Note that this is a manual cherry pick of
https://github.com/openshift/oc/pull/298/commits/9b68b418b0ee7a26b49c8d83b62fd788b81cce26
due to a merge conflict

https://bugzilla.redhat.com/show_bug.cgi?id=1804463