release: Allow a file:// release argument to release mirror
When a file:// argument is provided as the source for oc adm release
mirror, do a digest lookup within that repo for any image referenced
by the payload, and if found use the local directory as the source.
This allows releases to be mirrored to disk and then oc adm release
mirror to be used to push it back with:

  oc adm release mirror file://openshift/release:4.3.1 --to REPO

where before it would fail with "invalid image reference".

Ensure that when file:// is used the correct instructions are printed
for mirroring.

Fix the mirror command provided to users to put the wildcard in
quotes (to avoid bash shell expansion).