adm: Add a command that creates a token for a user
Creating a token for an admin or user is more difficult than it
has to be for scripting automation, and encourages use of service
account tokens instead. Add an administrative command that creates
a token under an oauth-client that may not be used for the standard
web flows for the current user or another.  Typical use would be:

    oc adm oauth create-token --user NAME

or if scoped tokens are desired, create a view token in the provided
namespace for the user.

    oc adm oauth create-token --user NAME --role=view -n NS

Admins can use this to get a token that is time bound but privileged.

In the future, we may wish to consider having an out of the box
client and admission control filter that allows a user to create
their own tokens with some constraints to simplify this flow.