Zeroing BearerTokenFile for prune
Kubernetes client-go requires only one authentication method to be present
therefore this patch zeroes the content of BearerTokenFile as basic auth
is used during prune.

In a nutshell: or we use token based authentication or basic auth, never
both at the same time.

This is a cherrypick of 966f1bb389f91503670b33263be35df346d9e568