login: choose the CAs based on the remote server cert
The previous check to help decide which CA to use when
performing a login is performed in such a way that fails in most
cases, causing the oauth-server to log `remote error: bad certificate`.

This commit fixes that by retrieving the server cert, closing the
connection and extracting the logic for certificate verification
locally without ruining the handshake with the server.