pkg/cli/admin/upgrade: Warn when --allow-upgrade-with-warnings let you through
To raise awareness that users are initiating or re-targeting an update
despite their cluster being unhealthy, which may help (updating to a
fixed release) or hurt (re-targeting a 4.1 -> 4.2 update to start
installing 4.3 components, which will not play nicely with any
remaining 4.1 components).  This is bit ugly due to the two
checkForUpgrade callsites, although I haven't tried to DRY that up.
And I've done some minimal shuffling so checkForUpgrade returns error
messages that are lowercased [1] and which can still be used in both
warning strings and fatal messages.

[1]: https://github.com/golang/go/wiki/CodeReviewComments#error-strings