pkg/cli/admin/upgrade: Delegate to update-approvers
This subcommand interacts with the cluster-version operator, which is
now maintained by the updates team.  I'm seeding the alias with the
content from [1] (and adding Lala, who's the team lead).  I've left
the 'component' property (added to the root OWNERS in e83f2ac9ae, Add
BZ component name to OWNERS, 2020-04-03, #367) off for now, because
the tooling around that only consumes entries at the repo level.

I think that the admin/release should also be maintained by the
updates team, because it manages the release images for which the CVO
is the sole consumer.  But Scott wants to punt on that for now.

[1]: https://github.com/openshift/cluster-version-operator/blob/914467c03b5bc25dbe4c9295c420fdb2598db821/OWNERS