debug: Allow debug to start with a single `--image` argument
The command `oc debug --image=foo` should be allowed to start foo
exactly as if you ran `oc debug istag/foo:latest`. This allows us
to have a minimal command for scenarios where you want to test the
image in the debug setup. The default image with no arguments for
`oc debug` is the openshift/tools image.