Bug 1799020: Add a push-config flag to 'oc adm release new'
It is not possible to configure 2 auths for a single image registry in a single authfile.
This is consistent with behavior of docker, podman, and oc.
Allow 'oc adm release new' to succeed when a user pulls from one quay.io repository that
requires one auth and then push to another quay.io repository that requires a different auth.

push-config defaults to ~/.docker/config.json, same as registry-config.