only print/update changed rolebindings when removing roles
The rolebinding modification code was attempting to update all roles
containing the specified role instead of just updating roles that need
updating.

The above would result in printing all rolebindings containing the
specified role if formatting flag was specified and could result
in errors being printed in case the formatting flag was not specified
but the specified role was contained in one of the system rolebindings.

Fixed by passing only the rolebindings that needed changing to the
updating code.