Add option to generate ImageContentSourcePolicy scoped to a registry
This enhancement adds --icsp-scope flag to `oc adm catalog mirror`. Scopes can be registry or
repository. Defaults to repository, which is the existing behavior. Widening the mirror
to registry scope in the icsp file minimizes the number of times the cluster needs to
reboot its nodes..

Implements https://github.com/openshift/enhancements/pull/374