cli: Add `oc adm prune imagestream` to prune image stream tags
This command is useful in circumstances where the image stream tags
must be pruned and general cluster pruning is not needed. Unlike
general image pruning references from other resources on the cluster
are not considered and so pruning image streams in this fashion may
result in other systems being unable to access these historical
images.

Accepts selectors and namespace constraints to filter multiple image
streams, and is dry-run by default. All tag revisions that would be
removed are printed in both modes.