prune: Use paging consistently with a higher chunk size by default
Images was using the default chunk size of 500 without the option
to change, but large clusters may not be able to serve all images
within the 5m window that a chunked list can offer. Use a higher
default and allow the default to be increased. Use chunking
consistently on all resource types to avoid a potential problem
with this in the future (for very large pod counts, for instance).