must-gather: use exponential backoff for polling
When running must-gather on a bare metal cluster, the command will fail
to find the pod after a 1 minute duration, and exit with an error
message in the form of `Get: $pod-url: unexpected EOF`.

This patch replaces the normal way of polling (every 1 minute), with an
exponential backoff starting with 1 second, and a factor of 2.  The
timeout value is unchanged.

Bug 1859972

Signed-off-by: Yuval Turgeman <yturgema@redhat.com>