mirror: Mirror layers which are only accessible via URL
A foreign image layer has a URL that points outside the current
registry, such that the registry returns not found if the blob
is requested. Since mirror attempts to ensure the destination
image can be pulled, even in the presence of disjoint networks,
and given that the Docker daemon will by default attempt the
blob lookup first before checking URLs, always copy the blob
by checking the source registry, then trying to pull the layer
URLs.

There may be scenarios where this is not desirable, and a future
change may allow skipping mirror of those blobs.