Bug 1812813: oc debug node: create debug namespace
When a defaultNodeSelector is configured in the scheduler, 'oc debug node' fails
if the node to debug does not have the matching labels.  To fix this, with
'oc debug node' create a temporary namespace for the debug pod that has an empty
'openshif.io/node-selector: ""' annotation.  Only create the debug ns if
 the user hasn't explicitly set the debug-pod namespace with
'--to-namespace' or '--namespace|-n'.