pkg/cli/admin/upgrade: Drop "previous" from --to's help text
We've had it at least since this repo was created in 3abf60a31e (setup
oc repo, 2019-06-17), but the --to implementation has only ever
iterated over availableUpdates.  This catches the help text up with
reality.

The confused help text is likely because cluster-version operator does
currently search history as well [1], but we're about to remove that
too, to make it harder to ask for dangerous rollbacks.

[1]: https://github.com/openshift/cluster-version-operator/blob/6d56c655ea16f6faee4b65ffef43dcd912657bc6/pkg/cvo/updatepayload.go#L305-L325