Reimplement image pruner
The new implementation of the image pruner does not use graphs. The
graph was replaced by two maps. The pruner works in two steps:

The first step deletes image stream items (aka revisions) that are not
used by the cluster components (pods, builds, etc.) and are not
protected by the pruner options (keep-younger-than, keep-tag-revisions,
etc.)

The second step deletes image objects that are not used by image
streams.