add remapping support for oc adm catalog mirror
the second step in performing a "true" airgapped install required
manual manipulation of the output mapping.txt and ICSP to properly
mirror.

The source of `oc adm catalog mirror` is always a catalog image that
has a certain set of image references contained within. When mirroring
these, an ImageContentSourcePolicy and mapping.txt file is generated.

The ICSP always maps the image reference in the catalog to the final
image defined by remapping into the destination registry, so that on
cluster the correct translation can take place.

The mapping.txt file (and the mirroring step performed by oc adm catalog
mirror) always need to map from an existing image to the target image.

When mirroring a catalog to an airgapped cluster, the images need to be
mirrored twice: once to the local filesystem, and once from the local
filesystem.

This commit adds support for a `--remap` flag which can be passed to
the mirror command. This "remaps" the source image found in the catalog
manifests to another location prior to mirroring, so that multi-stage
mirroring is easy.

For example:

mirror file://rh-ops:1 file://operators --to-manifests=local
mirror file://rh-ops:1 localhost:50000 --remap=./local/mapping.txt

This will mirror images from the external registry to the local fs, and
then again from the local fs to the final registry.

The ICSP generated is unmodified, since it will always need to map from
the images defined in the manifests themselves.