REVERT: Bug 1812813: oc debug node: create debug namespace with empty node-selector annotation #550
As discussed in the bugzilla report, creating the debug namespace that can override admin configured defaultNodeSelector gives too much privilege to oc.
We are reverting this change in favor of the original workaround, that is, for the admin to create a new debug namespace with the empty node-selector
and then direct oc to use that namespace for debugging.

$ oc adm new-project debug --node-selector=""
$ oc debug node/<node> -n debug