Trying out to address REFUSED_STREAM error
This is a naive approach to fix the REFUSED_STREAM problem we are seeing
in our CI cluster. It is not, by any means, a definitive solution as to
cache the blobs locally before copying them may not be an acceptable
solution.

If this address the problem successfuly we can then think on how to
properly fix the issue upstream.