idle: also annotate the service
Annotate the service with idle information as we currently do for
endpoints.

We choose to additionally annotate the service because the router in
4.6 has moved to watching endpointslices only. We don't annotate
endpointslices as they are managed by the endpointslice controller and
that doesn't propgagate annotations. The router now only checks the
service to see if it has been idled.

Previous discussion for propgagating annotations in endpointslices:

    https://github.com/kubernetes/kubernetes/pull/98379