vendor: Remove dependencies on older docker client code and bump
Move to a newer version fsouza/go-dockerclient that removes the
very, very, very, very old dependencies in our tree. At this point
only newapp uses the Docker API, but we want to remain as compatible
as possible. It is likely this breaks compatibility with RHEL7
clients using older Docker versions and that is ok.