pkg/cli/admin/inspect/namespace: Gather PDBs too
Pod disruption budgets are important enough to be included in Insights
tarballs [1], and they aren't all that big, so we should pick them up
as a standard part of interesting namespaces too.  This will help with
things like auditing issues where we failed to drain (PDB too strict)
or disrupted a core workload (PDB too weak or missing).

[1]: https://github.com/openshift/insights-operator/pull/185