when mirroring to a file destination, mount images under the index
when mirroring an index image: quay.io/my/index:a
that contains a reference to another image: quay.io/other/bundle:1
to a local file: file:///local/index

the referenced images are now mounted under the location for the index
itself, resulting in (for example):

file:///local/index/my/index:a
file:///local/index/my/index/other/bundle:1

this fixes a bug where the next step of an airgap mirror can't find the
images to be mirrored.

as a side effect, this keeps all catalog content stored under the same
fs location. Local file storage can be shared with other mirrored
images while keeping it simple to copy the full contents of an index
to another location.