release: oc adm release new should look at image stream status tags
The code that looks at spec tags when building a new release is
incorrectly assuming that a spec tag trumps the status tag, when in
fact the spec tag generation must be compared to the status tag
generation. Rearrange the code so that spec tags are skipped if the
spec tag is older than the status tag, but still errors if we are
waiting for a spec tag to be imported.

This may change behavior of OpenShift release creation in developer
environments, but the actual release tooling will now match the
behavior of the release. We were not seeing this bug impact us in
the release controller because the release controller clones the
image stream prior to creating a release, so spec/status are always
equal. This currently blocks users from creating releases off ocp/4.8
because there are spec tags that are old enough that their image has
been garbage collected in testing.