release: extract Linux binaries for multiple architectures
Previously, one could extract x86_64 Windows and Mac binaries (the only
architectures for those OSes until recently, and still the only ones
fully supported by Go at the moment) and the Linux binary for the
payload architecture, but there was no way to e.g. extract an x86_64
Linux binary to use as a client with a server of a different another
architecture.

This fixes that, and also adds the ability to extract an x86_64 Linux
installer from a payload of any architecture, on the basis that a remote
install should only require what the user has on their desk, and they
may not have (and should not need) a matching system for that purpose.