Allow users to specify max ICSP size
Problem: It is possible for the `oc adm catalog mirror` command
to generate ICSPs that are greater than 262144 bytes in size.
ICSPs that exceed 262144 bytes are likely to fail when applied
to the cluster if the objec already existed in an early state as
the `kubectl.kubernetes.io/last-applied-configuration` annotation
will likely exceed the 262144 annotation byte limit.

Solution: Introduce a the max-icsp-size flag to the
`oc adm catalog mirror` command, allowing users to specify the
maximum byte size of ICSP files generated by the command. If
an ICSP would exceed this limit, create and begin writting mirrors
to a new ICSP.

The default ICSP limit is 250000 bytes.