Download must-gather pod logs to get stdout info
Network-tools image leverages the must-gather tool
using the "--image" option. However some of the scripts
and tools used there just print the output to stdout.
It would be nice to get the must-gather pod logs as
well downloaded locally to see what happened while
the data was gathered or scripts were being run.