[must-gather]: ability to run must-gather on a set of nodes
It could be interesting for users to run must-gather pods simoultanously
on a set of cluster nodes. A typical use case for this comes from
networking and running `tcpdump` on a set of cluster nodes, allowing the
user to much more simply capture required packet data in a single
command. `--node-selector` also solves the use case where you as
a user: want to specify a command that can only be executed on linux /
windows nodes. This is particularily true in the case of `tcpdump`.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>