pkg/cli/admin/upgrade/upgrade: Copy edits, including "assists with cluster upgrades"
The previous:

  This command will request that the cluster begin an upgrade...

and similar confused some readers [1], because the no-argument
invocation is just a read query that does not request an update.  This
commit makes the wording more generic, to match the subcommand's
actual generic functionality (reading from and optionally writing to
ClusterVersion).

I've also tweaked the "No updates available" message to avoid folks
misreading as:

  ... doing so may not... result in downtime or data loss.

when we are trying to convey that updating despite a lack of
recommendations might be dangerous.

In both cases, the wording I'm altering dates back to the creation of
this subcommand in cd30f2f864 (Add `oc adm upgrade` to display
available updates or trigger an update, 2018-12-04,
openshift/origin#21605).

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1992680#c2