pkg/cli/admin/upgrade: Report on Failing!=False conditions
cae0b5e096 (React to degraded condition change, 2019-04-23,
openshift/origin#22644) moved this code from Failing to Degraded,
likely inspired by [1].  But Degraded is only used in ClusterOperator.
ClusterVersion kept using Failing, as seen in [2].  This commit
returns us to watching for Failing (the condition the CVO has been
setting the whole time), and informing the caller for any non-happy
statuses (or the lack of a Failing condition at all).

Even though the issue causing `Failing=True` may block the current
update from progressing, it should not block admins from requesting a
new update target.  For some bugs, retargeting is the recommended way
to resolve the issue that is currently sticking the update [3].

[1]: https://github.com/openshift/api/pull/287
[2]: https://github.com/openshift/cluster-version-operator/pull/191
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1988576#c30