Bug 1970805: Replace slashes in suggested ImageStream name
When figuring out an image stream name for created build objects,
SuggestName function can return name with slashes in some situations.

It is important to hold correct data (that including slashes, too) in
the ImageRef object, because it is used in other places and can lead to
errors.

SuggestName method of the ImageRef however, should return name
without slashes when suggesting name for an ImageStream.