[must-gather] Add options for multi-node / host-network capture (#850)
* [must-gather]: ability to run must-gather pods as hostNetwork: true

This can allow users specifying specific images for the must-gather, to
run commands requiring host-level access. In networking this can, among
other things, be necessary for packet captures or firmware information.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

[must-gather]: ability to run must-gather on a set of nodes

It could be interesting for users to run must-gather pods simoultanously
on a set of cluster nodes. A typical use case for this comes from
networking and running `tcpdump` on a set of cluster nodes, allowing the
user to much more simply capture required packet data in a single
command. `--node-selector` also solves the use case where you as
a user: want to specify a command that can only be executed on linux /
windows nodes. This is particularily true in the case of `tcpdump`.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

[must-gather]: expose node/pod specific information to must-gather commands

This could be useful for users which specify `--node-selector` and
capture must-gather data across multiple nodes, allowing them to
separate the captured and downloaded data between nodes.

Signed-off-by: Alexander Constantinescu <aconstan@redhat.com>

* Code review update