Make sure to set the namespace for routes resources (#881)
* Make sure to set the namespace for routes resources

When doing dry run, the namespace will not be part of the output
manifests ( i.e. yaml or json, etc. ). This commit adds the namespace
to the generated route resource if it is not the default.

* Set namespace for routes

Move the logic to enforce setting the namespace to UnsecuredRoute,
as it will apply to all invocations.