pkg/cli/admin/upgrade: Rename findCondition()
Rename the function findCondition to findClusterOperatorStatusCondition.

The function searches in a parameter of type
[]configv1.ClusterOperatorStatusCondition and returns
a type of *configv1.ClusterOperatorStatusCondition.

Rename the function to be more explicit about the types of its
parameters and its return type.
There's also a need to implement a similar function
for the type metav1.Condition, and there's a need to
differentiate between these functions.