pkg/cli/admin/upgrade: Modify the printed output of 'oc adm upgrade'
Modify the output of the command "oc adm upgrade".
Change "Available updates" to "Recommended updates".
Indent the printed versions.
Add a new line before the Recommended updates block for an esthetic
purpose.

Change the order of the printed versions to a decreasing order.
Since the sorting function is an internal helper function with only a
one use-case, I have modified the function to sort in decreasing order.

Increase the space between the "VERSION" and "IMAGE" columns.
Set their minimal cell width to 14 to have a larger space between
the columns for shorter versions. But for longer versions, there
won't be an unnecessary big empty space.

The changes are specified or inspired by the proposed enhancement for
Targeted Update Edge Blocking (https://github.com/openshift/enhancements/blob/master/enhancements/update/targeted-update-edge-blocking.md)
and its examples.
The decreasing order is specified in the proposed enhancement.
In the enhancement it is mentioned that "Both availableUpdates and
conditionalUpdates should be sorted in decreasing SemVer order for
stability, to avoid unnecessary status churn." The decreasing order
is also shown in the examples of using the "oc adm upgrade" command
when printing available updates.