pkg/cli/admin/upgrade: Display not-recommended versions
Display a prompt for not-recommended versions if they are available and
the user uses the "oc adm upgrade" command.

Display the not-recommended versions in a formatted decreasing order if
the user uses the "oc adm upgrade --include-not-recommended" command,
and not-recommended versions are available.

The formatting of the output is specified or inspired by the proposed
enhancement for Targeted Update Edge Blocking (https://github.com/openshift/enhancements/blob/master/enhancements/update/targeted-update-edge-blocking.md)
and its examples.
The decreasing order is specified in the proposed enhancement.
In the enhancement it is mentioned that "Both availableUpdates and
conditionalUpdates should be sorted in decreasing SemVer order for
stability, to avoid unnecessary status churn." The decreasing order is
also shown in the examples of using the "oc adm upgrade" when printing
not-recommended updates.