Bug 2039491: Remove `git://` from new-app tests
GitHub is removing support for the unsecured "git://" protocol when
cloning source code [1]. This is being done in phases, with full
removal happening on March 15, 2022. Some of the new-app unit tests
connect to GitHub and attempt to clone source code - this updates the
tests to use HTTPS instead of `git://`.

[1] https://github.blog/2021-09-01-improving-git-protocol-security-github/