Added login flow using a browser
This change adds a new flag called `--web` which launches the system
browser with the authorization endpoint as the target. The user can then
authenticate and is redirected back to a server which is listening on
the loopback address on a random port. The callback receives the
authorization code which it can exchange for a token and configure the
client.