pkg/cli/admin/upgrade: --to-image --allow-explicit-upgrade and known update targets
When folks use --to-image and --allow-explicit-upgrade together, oc
searches through recommended (and, with --allow-explicit-upgrade, also
supported but not recommended) targets to see if it knew about the
requested target image.

Until this commit, it would throw out that information, create an
Update that just pointed at the target release image, and warn the
user about the risks of --allow-explicit-upgrade.

With this commit, --allow-explicit-upgrade is consumed conditionally.
If we need it, because the earlier image search was unable to turn up
the requested release in ClusterVersion status, we'll still create an
Update and warn about the risks of --allow-explicit-upgrade.  But if
the search found the requested image (update != nil), we just use that
as if the user had requested it with --to, and we no longer pester the
user with complaints about the --allow-explicit-upgrade that they used
but, in this case, didn't need.