pkg/cli/admin/upgrade: Allow --to-lastest and --allow-not-recommended together
6c8d9352d3 (Adding the flag --allow-not-recommended to oc adm upgrade,
2021-12-01, #986) taught --to and --to-image about
--allow-not-recommended.  This commit extends that coverage to
--to-latest, allowing folks to say "I want the latest supported
update, even if that update is not currently recommended".  Probably
not the best idea to do that sort of thing, but that's generally true
of --allow-not-recommended.

To accomplish this change, I've consolidated into a single "user
requested an update" case, where we use one of the three options
(--to-latest, --to, or --to-image) to hunt for a target update, and
then centralized logic to actually apply that target.